pull SelectiveRetirementInput into separate file
