+++ app/components/views/Offset/styles.ts
@@ -189,53 +189,3 @@ export const modalContent = css`
     }
   }
 `;
-
-export const advancedButton = css`
-  width: 100%;
-  display: flex;
-  align-items: center;
-
-  &,
-  & .advancedButton_label {
-    color: var(--font-03);
-  }
-`;
-
-export const advancedButtonInput = css`
-  width: 100%;
-  display: flex;
-  align-items: center;
-  gap: 0.8rem;
-
-  input[data-error="true"] {
-    border: 0.2rem solid red;
-  }
-
-  .advancedButtonInput_iconAligner {
-    display: flex;
-    width: 100%;
-    position: relative;
-    align-items: center;
-  }
-  .deletebutton {
-    position: absolute;
-    right: 0.2rem;
-    background-color: var(--surface-02);
-  }
-  .plusbutton,
-  .deletebutton {
-    color: var(--font-03);
-    width: 3.4rem;
-    height: 3.4rem;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-  }
-  .plusbutton:hover,
-  .deletebutton:hover {
-    color: var(--font-01);
-  }
-  .plusbutton.hidden {
-    visibility: hidden;
-  }
-`;
